+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/iso/ZkSignedList.kt
@@ -1,0 +1,5 @@
+package at.asitplus.iso
+
+data class ZkSignedList(
+    val entries: List<ZkSignedItem>,
+)