+++ vck-openid-ktor/src/commonTest/kotlin/at/asitplus/wallet/lib/ktor/openid/OpenId4VpWalletTest.kt
@@ -3,6 +3,7 @@ package at.asitplus.wallet.lib.ktor.openid
 import at.asitplus.data.NonEmptyList.Companion.nonEmptyListOf
 import at.asitplus.dcapi.request.Oid4vpDCAPIRequest
 import at.asitplus.iso.IssuerSignedItem
+import at.asitplus.iso.Item
 import at.asitplus.openid.CredentialFormatEnum
 import at.asitplus.openid.OidcUserInfo
 import at.asitplus.openid.OidcUserInfoExtended
@@ -467,7 +468,7 @@ private fun SuccessIso.containsAttribute(attribute: Map.Entry<String, String>): 
 private fun SelectiveDisclosureItem.matchesAttribute(attribute: Map.Entry<String, String>): Boolean =
     claimName == attribute.key && claimValue.jsonPrimitive.content == attribute.value
 
-private fun IssuerSignedItem.matchesAttribute(attribute: Map.Entry<String, String>): Boolean =
+private fun Item.matchesAttribute(attribute: Map.Entry<String, String>): Boolean =
     elementIdentifier == attribute.key && elementValue.toString() == attribute.value
 
 // If the countdownLatch has been unlocked, the correct credential has been posted to the RP, and we're done!