+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/IsoPresentationMeta.kt
@@ -1,0 +1,9 @@
+package at.asitplus.wallet.lib.agent
+
+import at.asitplus.iso.ZkRequest
+import at.asitplus.jsonpath.core.NormalizedJsonPath
+
+class IsoPresentationMeta(
+    val claims: Collection<NormalizedJsonPath>,
+    val spec: ZkRequest = ZkRequest.Default
+)