+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/VerifierAgent.kt
@@ -4,6 +4,7 @@ import at.asitplus.catchingUnwrapped
 import at.asitplus.iso.DeviceResponse
 import at.asitplus.iso.Document
 import at.asitplus.iso.MobileSecurityObject
+import at.asitplus.iso.ZkDocument
 import at.asitplus.openid.TransactionDataBase64Url
 import at.asitplus.signum.indispensable.josef.JwsSigned
 import at.asitplus.wallet.lib.agent.Verifier.VerifyPresentationResult
@@ -46,9 +47,10 @@ class VerifierAgent(
 
     override suspend fun verifyPresentationIsoMdoc(
         input: DeviceResponse,
-        verifyDocument: suspend (MobileSecurityObject, Document) -> Boolean,
+        verifyPlainDocument: suspend (MobileSecurityObject, Document) -> Boolean,
+        verifyZkDocument: (ZkDocument) -> Boolean,
     ): VerifyPresentationResult = catchingUnwrapped {
-        validatorMdoc.verifyDeviceResponse(input, verifyDocument)
+        validatorMdoc.verifyDeviceResponse(input, verifyPlainDocument, verifyZkDocument)
     }.getOrElse {
         VerifyPresentationResult.ValidationError(it)
     }